refactor: split oversized TUI settings modules to meet 50-line budget
CodeTether-Provenance-ID: ctprov_17ca7fd88d09d92f92dac1bccd71bdd0
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: 0e65c04e-72b1-4cfa-b596-afaaff2c7155
